+++ ChangeLog.md
@@ -5,6 +5,8 @@
 Upcoming Release
 
 - Fixed a bug that to return the list of containers in sorted order.
+- Fixed a bug that get/download blob snapshot fail.
+- Check input request "x-ms-version" Header, only valid version are allowed
 
 Blob:
 - Added support for CORS.